requirements.txt
typer==0.4.1
colorama==0.4.3
shellingham==1.4.0
pytest==7.1.1
pyyaml==6.0